# QuickShip_SaaS_V4_Enterprise_Engineering_Handbook/docs/08_api/mobile_phase1_openapi.yaml
openapi: 3.1.0
info:
  title: Mathukai Operations Mobile API
  version: 1.0.0-draft
  summary: Phase 1 API contract for the Android mobile application.
  description: |
    Design contract for Mathukai Operations Phase 1. Django remains the source
    of truth for tenant access, permissions, order transitions, stock values,
    audit logging, and integration side effects. Barcode scanning and packing
    endpoints are intentionally excluded until Phase 2.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Same-origin API; deployment host is environment-specific.
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [Orders]
      operationId: listOrders
      summary: List tenant-scoped orders using cursor pagination.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: search
          in: query
          schema:
            type: string
            maxLength: 160
          description: Search permitted order, customer, or tracking fields.
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatusCode'
        - name: payment_state
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStateCode'
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: updated_after
          in: query
          schema:
            type: string
            format: date-time
          description: Supports foreground synchronization of changed records.
      responses:
        '200':
          description: Order page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionError'
  /orders/{order_id}:
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Return a tenant-scoped order and its currently allowed actions.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Ten-minute access token scoped to the selected tenant.
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    OrderId:
      name: order_id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    ValidationError:
      description: Request data is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_error
              message: Check the highlighted fields.
              fields:
                target_status: [This field is required.]
              retryable: false
            meta:
              request_id: req_01JZEXAMPLE
              server_time: '2026-07-19T10:00:00Z'
    AuthenticationError:
      description: Authentication is missing, expired, or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: authentication_required
              message: Sign in again to continue.
              fields: {}
              retryable: false
            meta:
              request_id: req_01JZEXAMPLE
              server_time: '2026-07-19T10:00:00Z'
    PermissionError:
      description: The current user or tenant cannot perform the operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: The resource is unavailable in the active tenant scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ResponseMeta:
      type: object
      required: [request_id, server_time]
      properties:
        request_id:
          type: string
        server_time:
          type: string
          format: date-time
        cache_expires_at:
          type: [string, 'null']
          format: date-time
    Pagination:
      type: object
      additionalProperties: false
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, fields, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            retryable:
              type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    OrderStatusCode:
      type: string
      enum:
        - new_order
        - order_accepted
        - order_packed
        - shipped
        - delivery_issue
        - out_for_delivery
        - delivered
        - completed
        - order_cancelled
    PaymentStateCode:
      type: string
      enum: [pending, received]
    LabeledStatus:
      type: object
      additionalProperties: false
      required: [code, label]
      properties:
        code:
          type: string
        label:
          type: string
    Money:
      type: object
      additionalProperties: false
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?[0-9]+\.[0-9]{2}$'
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: INR
    OrderSummary:
      type: object
      required:
        - id
        - reference
        - source
        - status
        - payment_state
        - item_count
        - total
        - order_date
        - attention_required
        - version
        - updated_at
      properties:
        id:
          type: integer
        reference:
          type: string
        source:
          $ref: '#/components/schemas/LabeledStatus'
        status:
          $ref: '#/components/schemas/LabeledStatus'
        payment_state:
          $ref: '#/components/schemas/LabeledStatus'
        customer_display_name:
          type: [string, 'null']
          description: Masked or omitted according to field-level permissions.
        item_count:
          type: integer
          minimum: 0
        total:
          $ref: '#/components/schemas/Money'
        order_date:
          type: [string, 'null']
          format: date-time
        tracking_number:
          type: [string, 'null']
        attention_required:
          type: boolean
        version:
          type: string
          description: Opaque concurrency token derived by the server.
        updated_at:
          type: string
          format: date-time
    CustomerSummary:
      type: object
      additionalProperties: false
      required: [name, phone, email, delivery_address]
      properties:
        name:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        email:
          type: [string, 'null']
          format: email
        delivery_address:
          type: [string, 'null']
        fields_masked:
          type: array
          items:
            type: string
    OrderItem:
      type: object
      additionalProperties: false
      required: [name, quantity, total]
      properties:
        product_id:
          type: [integer, 'null']
        name:
          type: string
        sku:
          type: [string, 'null']
        quantity:
          type: integer
          minimum: 1
        total:
          $ref: '#/components/schemas/Money'
        image_url:
          type: [string, 'null']
          format: uri
    AllowedAction:
      type: object
      additionalProperties: false
      required: [code, label, confirmation_required]
      properties:
        code:
          type: string
          enum: [update_status, mark_payment_received]
        label:
          type: string
        target_status:
          oneOf:
            - $ref: '#/components/schemas/OrderStatusCode'
            - type: 'null'
        confirmation_required:
          type: boolean
        reason_required:
          type: boolean
          default: false
        required_fields:
          type: array
          items:
            type: string
    OrderActivity:
      type: object
      additionalProperties: false
      required: [id, title, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: [string, 'null']
        actor_display_name:
          type: [string, 'null']
        previous_status:
          type: [string, 'null']
        current_status:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
    OrderDetail:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          required: [customer, items, courier_name, shipping_cost, shipping_gst, shipping_total, allowed_actions, activity]
          properties:
            customer:
              $ref: '#/components/schemas/CustomerSummary'
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            courier_name:
              type: [string, 'null']
            shipping_cost:
              $ref: '#/components/schemas/Money'
              description: Shipping charge before GST.
            shipping_gst:
              $ref: '#/components/schemas/Money'
              description: GST calculated at 18% of the shipping charge.
            shipping_total:
              $ref: '#/components/schemas/Money'
              description: Shipping charge plus 18% GST.
            payment_received_at:
              type: [string, 'null']
              format: date-time
            cancellation_reason:
              type: [string, 'null']
            cancellation_note:
              type: [string, 'null']
            allowed_actions:
              type: array
              items:
                $ref: '#/components/schemas/AllowedAction'
            activity:
              type: array
              items:
                $ref: '#/components/schemas/OrderActivity'
    OrderListResponse:
      type: object
      additionalProperties: false
      required: [data, pagination, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/OrderSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    OrderDetailResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/OrderDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...